feat(usage): extract Codex reset credit redemption hook
